rentspace/requirements.txt
asgiref==3.10.0
async-timeout==5.0.1
boto3==1.40.50
botocore==1.40.50
channels==4.3.1
channels_redis==4.3.0
Django==5.2.7
django-cors-headers==4.9.0
django-filter==25.2
django-storages==1.14.6
djangorestframework==3.16.1
djangorestframework-gis==1.2.0
djangorestframework_simplejwt==5.5.1
gunicorn==24.1.1
jmespath==1.0.1
msgpack==1.1.2
packaging==26.0
pillow==11.3.0
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
redis==6.4.0
s3transfer==0.14.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.15.0
urllib3==2.5.0
whitenoise==6.11.0
